Tests for tools iroha_migrate, iroha_wsv_diff (#1597)
* Commit WSV every 1000 blocks during reindex
Closes #1339
* [tests] iroha-migrate
* test_tool_wsv_checker.sh
* improve test wsv_checker
* Rename iroha_migrate and iroha_wsv_diff
* fix iceseer's comments, not all
* Block store from Stepan
* adapt to another blockchain
* Fix lebdron tests
* cross-platform
* macos bash version 5
* on pull_request 'build all'
* CTEST_module_timeout=120
