+++ irohad/CMakeLists.txt
@@ -17,5 +17,5 @@ add_subdirectory(multi_sig_transactions)
 add_subdirectory(pending_txs_storage)
 add_subdirectory(util)
 add_subdirectory(maintenance)
-add_subdirectory(migration-tool)
-add_subdirectory(wsv_checker)
+add_subdirectory(iroha_migrate)
+add_subdirectory(iroha_wsv_diff)
