+++ irohad/ametsuchi/impl/wsv_restorer_impl.cpp
@@ -252,6 +252,8 @@ namespace iroha::ametsuchi {
           break;
       } while (1);
 
+      /// Periodically poll if there new blocks in block_storage
+      /// (on filesystem). When more blocks found, reindex them.
       while (wait_for_new_blocks) {
         std::this_thread::sleep_for(kWaitForBlockTime);
         block_query->reloadBlockstore();
@@ -262,7 +264,7 @@ namespace iroha::ametsuchi {
         while (hasError(block_result)
                && (new_last_block > last_block_in_storage)) {
           --new_last_block;
-          auto block_result = block_query->getBlock(new_last_block);
+          block_result = block_query->getBlock(new_last_block);
         };
 
         if (new_last_block > last_block_in_storage) {
