+++ test/framework/integration_framework/fake_peer/network/on_demand_os_network_notifier.cpp
@@ -56,6 +56,10 @@ namespace integration_framework::fake_peer {
     return true;
   }
 
+  bool OnDemandOsNetworkNotifier::hasEnoughBatchesInCache() const {
+    return false;
+  }
+
   void OnDemandOsNetworkNotifier::processReceivedProposal(
       CollectionType batches) {}
