+++ test/module/irohad/ordering/on_demand_os_server_grpc_test.cpp
@@ -134,7 +134,7 @@ TEST_F(OnDemandOsServerGrpcTest, RequestProposal) {
       std::make_shared<const shared_model::proto::Proposal>(proposal));
   EXPECT_CALL(*notification, onRequestProposal(round))
       .WillOnce(Return(ByMove(std::move(iproposal))));
-  EXPECT_CALL(*notification, isEmptyBatchesCache()).WillOnce(Return(false));
+  EXPECT_CALL(*notification, hasEnoughBatchesInCache()).WillOnce(Return(true));
 
   grpc::ServerContext context;
   server->RequestProposal(&context, &request, &response);
@@ -162,7 +162,7 @@ TEST_F(OnDemandOsServerGrpcTest, RequestProposalNone) {
   proto::ProposalResponse response;
   EXPECT_CALL(*notification, onRequestProposal(round))
       .WillOnce(Return(ByMove(std::move(std::nullopt))));
-  EXPECT_CALL(*notification, isEmptyBatchesCache()).WillOnce(Return(true));
+  EXPECT_CALL(*notification, hasEnoughBatchesInCache()).WillOnce(Return(false));
 
   grpc::ServerContext context;
   server->RequestProposal(&context, &request, &response);
