+++ test/tool/irohad.restore_wsv.config
@@ -1,0 +1,14 @@
+{
+  "pg_opt": "dbname=iroha_data host=localhost port=5432 user=postgres password=postgres",
+  "block_store_path": "/PATH/TO/iroha/test/tool/block_store_7000",
+  "torii_port": 50051,
+  "internal_port": 10001,
+  "max_proposal_size": 10,
+  "proposal_delay": 5000,
+  "vote_delay": 5000,
+  "mst_enable": false,
+  "mst_expiration_time": 1440,
+  "max_rounds_delay": 3000,
+  "stale_stream_max_rounds": 2,
+  "metrics": "0.0.0.0:7001"
+}
